+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Unreleased
+- The config screen now uses [Cloth Config](https://www.curseforge.com/minecraft/mc-mods/cloth-config) (not required but only visible if installed)
 - Fixed invalid formatting in compiled `fabric.mod.json` ([#10](https://github.com/A5b84/dark-loading-screen/pull/10))
 
 ### [1.5.0]